requirements.txt
absl-py==0.9.0
appnope==0.1.2
asgiref==3.3.1
astor==0.8.1
astunparse==1.6.3
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cryptography==3.2.1
cycler==0.10.0
cytoolz==0.10.1
dask==2.30.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.3
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.34.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
image==1.5.32
imagecodecs==2020.5.30
imageio==2.9.0
importlib-metadata==3.1.1
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
joblib==0.17.0
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.3.3
matplotlib==3.3.3
networkx==2.4
numpy==1.19.4
oauthlib==3.0.1
olefile==0.46
opt-einsum==3.3.0
pandas==1.1.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.12.4
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pyexiv2==2.3.2
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-addons==0.10.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.11.26
toolz==0.10.0
tornado==6.0.4
tqdm==4.54.0
traitlets==4.3.3
typeguard==2.9.1
urllib3==1.26.2
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
